Rename `useDiscussionParams.ts`
